Remove usages of PolygonDrawingView and move reusable business logic to PolygonDrawingMapViewModel.kt
