+++ ground/src/main/res/layout/map_container_frag.xml
@@ -26,45 +26,34 @@
     <variable
       name="homeScreenViewModel"
       type="com.google.android.ground.ui.home.HomeScreenViewModel" />
-
-    <variable
-      name="polygonDrawingViewModel"
-      type="com.google.android.ground.ui.home.mapcontainer.PolygonDrawingViewModel" />
   </data>
 
   <FrameLayout
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <FrameLayout
-      android:id="@+id/map_overlay"
-      android:layout_width="match_parent"
-      android:layout_height="match_parent">
+    <include
+      android:id="@+id/basemap"
+      layout="@layout/basemap_layout"
+      app:viewModel="@{viewModel}" />
 
-      <include
-        android:id="@+id/basemap"
-        layout="@layout/basemap_layout"
-        android:visibility="@{viewModel.mapControlsVisibility}"
-        app:viewModel="@{viewModel}" />
+    <com.google.android.material.card.MaterialCardView
+      android:layout_width="wrap_content"
+      android:layout_height="wrap_content"
+      android:layout_gravity="center|bottom"
+      android:layout_margin="10dp">
 
-      <com.google.android.material.card.MaterialCardView
-        android:layout_width="wrap_content"
+      <TextView
+        android:id="@+id/location_accuracy"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_gravity="center|bottom"
-        android:layout_margin="10dp">
-
-        <TextView
-          android:id="@+id/location_accuracy"
-          android:layout_width="match_parent"
-          android:layout_height="wrap_content"
-          android:background="@android:color/white"
-          android:gravity="center"
-          android:padding="5dp"
-          android:text="@{viewModel.getLocationAccuracy()}"
-          android:textColor="@color/colorPrimary"
-          app:visible="@{viewModel.isLocationUpdatesEnabled()}" />
-      </com.google.android.material.card.MaterialCardView>
-    </FrameLayout>
+        android:background="@android:color/white"
+        android:gravity="center"
+        android:padding="5dp"
+        android:text="@{viewModel.getLocationAccuracy()}"
+        android:textColor="@color/colorPrimary"
+        app:visible="@{viewModel.isLocationUpdatesEnabled()}" />
+    </com.google.android.material.card.MaterialCardView>
 
     <ImageView
       android:id="@+id/map_crosshairs_img"
